refactor(tests): mover managers/ e consolidar 3 arquivos PostgresManager
Merge test_postgres_manager.py, test_postgres_manager_features.py e
test_postgres_manager_typesense.py em um único arquivo (58 testes).
Cria managers/conftest.py com fixtures compartilhadas (pg, mock_conn,
mock_dataset_manager_*). Move mock_dataset_manager_* do conftest raiz,
simplificando-o para conter apenas set_test_environment.

Refs: #136
