+++ tests/unit/managers/test_storage_adapter.py
